@@ -17,19 +17,36 @@ from __future__ import annotations
 
 import colorsys
 import typing
+from dataclasses import dataclass
 from warnings import warn
 
 import numpy as np
-import pandas as pd
 
 from .bounds import rescale
-from .external import crayon_rgb, husl, xkcd_rgb
+from .colors import (
+    CubeHelixMap,
+    GradientMap,
+    get_colormap,
+    get_named_color,
+    hex_to_rgb,
+    hsluv,
+    rgb_to_hex,
+)
 from .utils import identity
 
 if typing.TYPE_CHECKING:
-    from typing import Literal
+    from typing import Any, Literal, Optional, Sequence
 
-    from mizani.colors.typing import ColorScheme, ColorSchemeShort
+    from mizani.typing import (
+        Callable,
+        ColorScheme,
+        ColorSchemeShort,
+        FloatVector,
+        NumVector,
+        RGBHexColor,
+        TupleFloat2,
+        TupleFloat3,
+    )
 
 
 __all__ = [
@@ -52,7 +69,33 @@ __all__ = [
 ]
 
 
-def hls_palette(n_colors=6, h=0.01, l=0.6, s=0.65):
+class _discrete_pal:
+    """
+    Discrete palette maker
+    """
+
+    def __call__(self, n: int) -> Sequence[Any]:
+        """
+        Palette method
+        """
+        ...
+
+
+class _continuous_pal:
+    """
+    Continuous palette maker
+    """
+
+    def __call__(self, x: Sequence[Any]) -> FloatVector:
+        """
+        Palette method
+        """
+        ...
+
+
+def hls_palette(
+    n_colors: int = 6, h: float = 0.01, l: float = 0.6, s: float = 0.65
+) -> Sequence[TupleFloat3]:
     """
     Get a set of evenly spaced colors in HLS hue space.
 
@@ -95,7 +138,9 @@ def hls_palette(n_colors=6, h=0.01, l=0.6, s=0.65):
     return palette
 
 
-def husl_palette(n_colors=6, h=0.01, s=0.9, l=0.65):
+def husl_palette(
+    n_colors: int = 6, h: float = 0.01, s: float = 0.9, l: float = 0.65
+) -> Sequence[TupleFloat3]:
     """
     Get a set of evenly spaced colors in HUSL hue space.
 
@@ -136,11 +181,12 @@ def husl_palette(n_colors=6, h=0.01, s=0.9, l=0.65):
     hues *= 359
     s *= 99
     l *= 99
-    palette = [husl.husl_to_rgb(h_i, s, l) for h_i in hues]
+    palette = [hsluv.hsluv_to_rgb((h_i, s, l)) for h_i in hues]
     return palette
 
 
-def rescale_pal(range=(0.1, 1)):
+@dataclass
+class rescale_pal(_continuous_pal):
     """
     Rescale the input to the specific output range.
 
@@ -172,13 +218,14 @@ def rescale_pal(range=(0.1, 1)):
     array([0.1 , 0.1 , 0.28, 0.46, 0.82, 1.  , 1.  ])
     """
 
-    def _rescale(x):
-        return rescale(x, range, _from=(0, 1))
+    range: TupleFloat2 = (0.1, 1)
 
-    return _rescale
+    def __call__(self, x: FloatVector) -> FloatVector:
+        return rescale(x, self.range, _from=(0, 1))
 
 
-def area_pal(range=(1, 6)):
+@dataclass
+class area_pal(_continuous_pal):
     """
     Point area palette (continuous).
 
@@ -206,13 +253,14 @@ def area_pal(range=(1, 6)):
     area space, i.e it is squared.
     """
 
-    def area_palette(x):
-        return rescale(np.sqrt(x), to=range, _from=(0, 1))
+    range: TupleFloat2 = (1, 6)
 
-    return area_palette
+    def __call__(self, x: FloatVector) -> FloatVector:
+        return rescale(np.sqrt(x), to=self.range, _from=(0, 1))
 
 
-def abs_area(max):
+@dataclass
+class abs_area(_continuous_pal):
     """
     Point area palette (continuous), with area proportional to value.
 
@@ -241,13 +289,14 @@ def abs_area(max):
     maximum.
     """
 
-    def abs_area_palette(x):
-        return rescale(np.sqrt(np.abs(x)), to=(0, max), _from=(0, 1))
+    max: float
 
-    return abs_area_palette
+    def __call__(self, x: NumVector) -> NumVector:
+        return rescale(np.sqrt(np.abs(x)), to=(0, self.max), _from=(0, 1))
 
 
-def grey_pal(start=0.2, end=0.8):
+@dataclass
+class grey_pal(_discrete_pal):
     """
     Utility for creating continuous grey scale palette
 
@@ -269,27 +318,28 @@ def grey_pal(start=0.2, end=0.8):
     --------
     >>> palette = grey_pal()
     >>> palette(5)
-    ['#333333', '#737373', '#989898', '#b5b5b5', '#cccccc']
+    ['#333333', '#737373', '#989898', '#b4b4b4', '#cccccc']
     """
-    import matplotlib.colors as mcolors
 
-    gamma = 2.2
-    ends = ((0.0, start, start), (1.0, end, end))
-    cdict = {"red": ends, "green": ends, "blue": ends}
-    grey_cmap = mcolors.LinearSegmentedColormap("grey", cdict)
+    start: float = 0.2
+    end: float = 0.8
 
-    def continuous_grey_palette(n):
+    def __post_init__(self):
+        colors = (self.start,) * 3, (self.end,) * 3
+        self._gmap = GradientMap(colors)
+
+    def __call__(self, n: int) -> Sequence[RGBHexColor | None]:
+        gamma = 2.2
         # The grey scale points are linearly separated in
         # gamma encoded space
-        x = np.linspace(start**gamma, end**gamma, n)
+        space = np.linspace(self.start**gamma, self.end**gamma, n)
         # Map points onto the [0, 1] palette domain
-        vals = (x ** (1.0 / gamma) - start) / (end - start)
-        return ratios_to_colors(vals, grey_cmap)
-
-    return continuous_grey_palette
+        x = (space ** (1.0 / gamma) - self.start) / (self.end - self.start)
+        return self._gmap.continuous_palette(x)
 
 
-def hue_pal(h=0.01, l=0.6, s=0.65, color_space="hls"):
+@dataclass
+class hue_pal(_discrete_pal):
     """
     Utility for making hue palettes for color schemes.
 
@@ -321,34 +371,34 @@ def hue_pal(h=0.01, l=0.6, s=0.65, color_space="hls"):
     >>> hue_pal(color_space='husl')(5)
     ['#e0697e', '#9b9054', '#569d79', '#5b98ab', '#b675d7']
     """
-    import matplotlib.colors as mcolors
 
-    if not all(0 <= val <= 1 for val in (h, l, s)):
-        msg = (
-            "hue_pal expects values to be between 0 and 1. "
-            " I got h={}, l={}, s={}".format(h, l, s)
-        )
-        raise ValueError(msg)
+    h: float = 0.01
+    l: float = 0.6
+    s: float = 0.65
+    color_space: Literal["hls", "husl"] = "hls"
 
-    if color_space not in ("hls", "husl"):
-        msg = "color_space should be one of ['hls', 'husl']"
-        raise ValueError(msg)
+    def __post_init__(self):
+        h, l, s = self.h, self.l, self.s
+        if not all(0 <= val <= 1 for val in (h, l, s)):
+            msg = (
+                "hue_pal expects values to be between 0 and 1. "
+                f"I got {h=}, {l=}, {s=}"
+            )
+            raise ValueError(msg)
 
-    name = "{}_palette".format(color_space)
-    palette = globals()[name]
+        if self.color_space not in ("hls", "husl"):
+            msg = "color_space should be one of ['hls', 'husl']"
+            raise ValueError(msg)
 
-    def _hue_pal(n):
-        colors = palette(n, h=h, l=l, s=s)
-        return [mcolors.rgb2hex(c) for c in colors]
+    def __call__(self, n: int) -> Sequence[RGBHexColor]:
+        lookup = {"husl": husl_palette, "hls": hls_palette}
+        palette = lookup[self.color_space]
+        colors = palette(n, h=self.h, l=self.l, s=self.s)
+        return [hsluv.rgb_to_hex(c) for c in colors]
 
-    return _hue_pal
 
-
-def brewer_pal(
-    type: ColorScheme | ColorSchemeShort = "seq",
-    palette: int = 1,
-    direction: Literal[1, -1] = 1,
-):
+@dataclass
+class brewer_pal(_discrete_pal):
     """
     Utility for making a brewer palette
 
@@ -396,71 +446,39 @@ def brewer_pal(
         print(get_palette_names("qualitative"))
         print(get_palette_names("diverging"))
     """
-    from .colors import brewer
 
-    if direction != 1 and direction != -1:
-        raise ValueError("direction should be 1 or -1.")
+    type: ColorScheme | ColorSchemeShort = "seq"
+    palette: int = 1
+    direction: Literal[1, -1] = 1
 
-    pal = brewer.get_color_palette(type, palette)
+    def __post_init__(self):
+        from .colors import brewer
 
-    def _brewer_pal(n):
+        if self.direction not in (1, -1):
+            raise ValueError("direction should be 1 or -1.")
+
+        self.bpal = brewer.get_color_palette(self.type, self.palette)
+
+    def __call__(self, n: int) -> Sequence[RGBHexColor | None]:
         # Only draw the maximum allowable colors from the palette
         # and fill any remaining spots with None
-        _n = min(max(n, pal.min_colors), pal.max_colors)
-        color_map = pal.get_hex_swatch(_n)
+        _n = min(max(n, self.bpal.min_colors), self.bpal.max_colors)
+        color_map = self.bpal.get_hex_swatch(_n)
         colors = color_map[:n]
-        if n > pal.max_colors:
+        if n > self.bpal.max_colors:
             msg = (
                 "Warning message:"
-                f"Brewer palette {pal.name} has a maximum "
-                f"of {pal.max_colors} colors Returning the palette you "
-                "asked for with that many colors"
+                f"Brewer palette {self.bpal.name} has a maximum "
+                f"of {self.bpal.max_colors} colors Returning the "
+                "palette you asked for with that many colors"
             )
             warn(msg)
-            colors = colors + [None] * (n - pal.max_colors)
-        return colors[::direction]
-
-    return _brewer_pal
+            colors = colors + [None] * (n - self.bpal.max_colors)
+        return colors[:: self.direction]
 
 
-def ratios_to_colors(values, colormap):
-    """
-    Map values in the range [0, 1] onto colors
-
-    Parameters
-    ----------
-    values : array_like | float
-        Numeric(s) in the range [0, 1]
-    colormap : cmap
-        Matplotlib colormap to use for the mapping
-
-    Returns
-    -------
-    out : list | float
-        Color(s) corresponding to the values
-    """
-    import matplotlib.colors as mcolors
-
-    iterable = True
-    try:
-        iter(values)
-    except TypeError:
-        iterable = False
-        values = [values]
-
-    color_tuples = colormap(values)
-    hex_colors = [mcolors.rgb2hex(t) for t in color_tuples]
-
-    nan_bool_idx = pd.isnull(values) | np.isinf(values)
-    if any(nan_bool_idx):
-        hex_colors = [
-            np.nan if is_nan else color
-            for color, is_nan in zip(hex_colors, nan_bool_idx)
-        ]
-    return hex_colors if iterable else hex_colors[0]
-
-
-def gradient_n_pal(colors, values=None, name="gradientn"):
+@dataclass
+class gradient_n_pal(_continuous_pal):
     """
     Create a n color gradient palette
 
@@ -472,8 +490,6 @@ def gradient_n_pal(colors, values=None, name="gradientn"):
         list of points in the range [0, 1] at which to
         place each color. Must be the same size as
         `colors`. Default to evenly space the colors
-    name : str
-        Name to call the resultant MPL colormap
 
     Returns
     -------
@@ -488,39 +504,30 @@ def gradient_n_pal(colors, values=None, name="gradientn"):
     --------
     >>> palette = gradient_n_pal(['red', 'blue'])
     >>> palette([0, .25, .5, .75, 1])
-    ['#ff0000', '#bf0040', '#7f0080', '#3f00c0', '#0000ff']
+    ['#ff0000', '#bf0040', '#7f0080', '#4000bf', '#0000ff']
     >>> palette([-np.inf, 0, np.nan, 1, np.inf])
-    [nan, '#ff0000', nan, '#0000ff', nan]
+    [None, '#ff0000', None, '#0000ff', None]
     """
-    import matplotlib.colors as mcolors
 
-    # Note: For better results across devices and media types,
-    # it would be better to do the interpolation in
-    # Lab color space.
-    if values is None:
-        colormap = mcolors.LinearSegmentedColormap.from_list(name, colors)
-    else:
-        colormap = mcolors.LinearSegmentedColormap.from_list(
-            name, list(zip(values, colors))
-        )
+    colors: Sequence[str]
+    values: Optional[Sequence[float]] = None
 
-    def _gradient_n_pal(vals):
-        return ratios_to_colors(vals, colormap)
+    def __post_init__(self):
+        self._gmap = GradientMap(self.colors, self.values)
 
-    return _gradient_n_pal
+    def __call__(self, x: FloatVector) -> Sequence[RGBHexColor | None]:
+        return self._gmap.continuous_palette(x)
 
 
-def cmap_pal(name, lut=None):
+@dataclass
+class cmap_pal(_continuous_pal):
     """
-    Create a continuous palette using an MPL colormap
+    Create a continuous palette using a colormap
 
     Parameters
     ----------
     name : str
         Name of colormap
-    lut : None | int
-        This is the number of entries desired in the lookup table.
-        Default is ``None``, leave it up Matplotlib.
 
     Returns
     -------
@@ -537,38 +544,25 @@ def cmap_pal(name, lut=None):
     >>> palette([.1, .2, .3, .4, .5])
     ['#482475', '#414487', '#355f8d', '#2a788e', '#21918c']
     """
-    import matplotlib as mpl
 
-    colormap = mpl.colormaps[name]
+    name: str
 
-    if lut is not None:
-        warn(
-            "The lut parameter has been deprecated and will "
-            "be removed in a future version.",
-            FutureWarning,
-        )
-        resample = getattr(
-            colormap, "resampled", getattr(colormap, "_resample", None)
-        )
-        colormap = resample(lut)
+    def __post_init__(self):
+        self.cm = get_colormap(self.name)
 
-    def _cmap_pal(vals):
-        return ratios_to_colors(vals, colormap)
-
-    return _cmap_pal
+    def __call__(self, x: FloatVector) -> Sequence[RGBHexColor | None]:
+        return self.cm.continuous_palette(x)
 
 
-def cmap_d_pal(name, lut=None):
+@dataclass
+class cmap_d_pal(_discrete_pal):
     """
-    Create a discrete palette using an MPL Listed colormap
+    Create a discrete palette from a colormap
 
     Parameters
     ----------
     name : str
         Name of colormap
-    lut : None | int
-        This is the number of entries desired in the lookup table.
-        Default is ``None``, leave it up Matplotlib.
 
     Returns
     -------
@@ -582,58 +576,35 @@ def cmap_d_pal(name, lut=None):
     --------
     >>> palette = cmap_d_pal('viridis')
     >>> palette(5)
-    ['#440154', '#3b528b', '#21918c', '#5cc863', '#fde725']
+    ['#440154', '#3b528b', '#21918c', '#5ec962', '#fde725']
     """
-    import matplotlib as mpl
-    import matplotlib.colors as mcolors
 
-    colormap = mpl.colormaps[name]
+    name: str
 
-    if lut is not None:
-        warn(
-            "The lut parameter has been deprecated and will "
-            "be removed in a future version.",
-            FutureWarning,
-        )
-        resample = getattr(
-            colormap, "resampled", getattr(colormap, "_resample", None)
-        )
-        colormap = resample(lut)
+    def __post_init__(self):
+        self.cm = get_colormap(self.name)
 
-    if not isinstance(colormap, mcolors.ListedColormap):
-        raise ValueError(
-            "For a discrete palette, cmap must be of type "
-            "matplotlib.colors.ListedColormap"
-        )
+        # Currently all named colormaps have an array of colors
+        # We do not expect CubeHelixMap
+        self.num_colors = len(self.cm.colors)
 
-    ncolors = len(colormap.colors)
-
-    def _cmap_d_pal(n):
-        if n > ncolors:
+    def __call__(self, n: int) -> Sequence[RGBHexColor]:
+        if n > self.num_colors:
             raise ValueError(
-                "cmap `{}` has {} colors you requested {} "
-                "colors.".format(name, ncolors, n)
+                f"cmap {self.name} has {self.num_colors} colors you "
+                f"requested {n} colors."
             )
-
-        if ncolors < 256:
-            return [mcolors.rgb2hex(c) for c in colormap.colors[:n]]
-        else:
-            # Assume these are continuous and get colors equally spaced
-            # intervals  e.g. viridis is defined with 256 colors
-            idx = np.linspace(0, ncolors - 1, n).round().astype(int)
-            return [mcolors.rgb2hex(colormap.colors[i]) for i in idx]
-
-    return _cmap_d_pal
+        return self.cm.discrete_palette(n)
 
 
-def desaturate_pal(color, prop, reverse=False):
+class desaturate_pal(gradient_n_pal):
     """
     Create a palette that desaturate a color by some proportion
 
     Parameters
     ----------
-    color : matplotlib color
-        hex, rgb-tuple, or html color name
+    color : color
+        html color name, hex, rgb-tuple
     prop : float
         saturation channel of color will be multiplied by
         this value
@@ -655,26 +626,27 @@ def desaturate_pal(color, prop, reverse=False):
     >>> palette([0, .25, .5, .75, 1])
     ['#ff0000', '#e21d1d', '#c53a3a', '#a95656', '#8c7373']
     """
-    import matplotlib.colors as mcolors
 
-    if not 0 <= prop <= 1:
-        raise ValueError("prop must be between 0 and 1")
+    def __init__(self, color: str, prop: float, reverse: bool = False):
+        if not 0 <= prop <= 1:
+            raise ValueError("prop must be between 0 and 1")
 
-    # Get rgb tuple rep
-    # Convert to hls
-    # Desaturate the saturation channel
-    # Convert back to rgb
-    rgb = mcolors.colorConverter.to_rgb(color)
-    h, l, s = colorsys.rgb_to_hls(*rgb)
-    s *= prop
-    desaturated_color = colorsys.hls_to_rgb(h, l, s)
-    colors = [color, desaturated_color]
-    if reverse:
-        colors = colors[::-1]
-    return gradient_n_pal(colors, name="desaturated")
+        if isinstance(color, str):
+            color = get_named_color(color)
+
+        rgb = hex_to_rgb(color)
+        h, l, s = colorsys.rgb_to_hls(*rgb)
+        s *= prop
+        desaturated_color = rgb_to_hex(colorsys.hls_to_rgb(h, l, s))
+        colors = [color, desaturated_color]
+        if reverse:
+            colors = colors[::-1]
+
+        super().__init__(colors)
 
 
-def manual_pal(values):
+@dataclass
+class manual_pal(_discrete_pal):
     """
     Create a palette from a list of values
 
@@ -695,22 +667,22 @@ def manual_pal(values):
     >>> palette(3)
     ['a', 'b', 'c']
     """
-    max_n = len(values)
 
-    def _manual_pal(n):
-        if n > max_n:
-            msg = (
-                "Palette can return a maximum of {} values. "
-                "{} values requested."
+    values: Sequence[Any]
+
+    def __post_init__(self):
+        self.size = len(self.values)
+
+    def __call__(self, n: int) -> Sequence[Any]:
+        if n > self.size:
+            warn(
+                f"Palette can return a maximum of {self.size} values. "
+                f"{n} values requested."
             )
-            warn(msg.format(max_n, n))
-
-        return values[:n]
-
-    return _manual_pal
+        return self.values[:n]
 
 
-def xkcd_palette(colors):
+def xkcd_palette(colors: Sequence[str]) -> Sequence[RGBHexColor]:
     """
     Make a palette with color names from the xkcd color survey.
 
@@ -719,7 +691,7 @@ def xkcd_palette(colors):
     Parameters
     ----------
     colors : list of strings
-        List of keys in the ``mizani.external.xkcd_rgb`` dictionary.
+        List of keys in the ``mizani.colors.xkcd_rgb`` dictionary.
 
     Returns
     -------
@@ -730,16 +702,16 @@ def xkcd_palette(colors):
     --------
     >>> palette = xkcd_palette(['red', 'green', 'blue'])
     >>> palette
-    ['#e50000', '#15b01a', '#0343df']
+    ['#E50000', '#15B01A', '#0343DF']
 
-    >>> from mizani.external import xkcd_rgb
-    >>> list(sorted(xkcd_rgb.keys()))[:5]
-    ['acid green', 'adobe', 'algae', 'algae green', 'almost black']
+    >>> from mizani.colors.named_colors import XKCD
+    >>> list(sorted(XKCD.keys()))[:4]
+    ['xkcd:acid green', 'xkcd:adobe', 'xkcd:algae', 'xkcd:algae green']
     """
-    return [xkcd_rgb[name] for name in colors]
+    return [get_named_color(f"xkcd:{name}") for name in colors]
 
 
-def crayon_palette(colors):
+def crayon_palette(colors: Sequence[str]) -> Sequence[RGBHexColor]:
     """
     Make a palette with color names from Crayola crayons.
 
@@ -749,7 +721,7 @@ def crayon_palette(colors):
     Parameters
     ----------
     colors : list of strings
-        List of keys in the ``mizani.external.crayloax_rgb`` dictionary.
+        List of keys in the ``mizani.colors.crayloax_rgb`` dictionary.
 
     Returns
     -------
@@ -760,20 +732,19 @@ def crayon_palette(colors):
     --------
     >>> palette = crayon_palette(['almond', 'silver', 'yellow'])
     >>> palette
-    ['#eed9c4', '#c9c0bb', '#fbe870']
+    ['#EED9C4', '#C9C0BB', '#FBE870']
 
-    >>> from mizani.external import crayon_rgb
-    >>> list(sorted(crayon_rgb.keys()))[:5]
-    ['almond', 'antique brass', 'apricot', 'aquamarine', 'asparagus']
+    >>> from mizani.colors.named_colors import CRAYON
+    >>> list(sorted(CRAYON.keys()))[:3]
+    ['crayon:almond', 'crayon:antique brass', 'crayon:apricot']
     """
-    return [crayon_rgb[name] for name in colors]
+    return [get_named_color(f"crayon:{name}") for name in colors]
 
 
-def cubehelix_pal(
-    start=0, rot=0.4, gamma=1.0, hue=0.8, light=0.85, dark=0.15, reverse=False
-):
+@dataclass
+class cubehelix_pal(_discrete_pal):
     """
-    Utility for creating continuous palette from the cubehelix system.
+    Utility for creating discrete palette from the cubehelix system.
 
     This produces a colormap with linearly-decreasing (or increasing)
     brightness. That means that information will be preserved if printed to
@@ -815,22 +786,33 @@ def cubehelix_pal(
     --------
     >>> palette = cubehelix_pal()
     >>> palette(5)
-    ['#edd1cb', '#d499a7', '#aa688f', '#6e4071', '#2d1e3e']
+    ['#edd1cb', '#d499a7', '#aa678f', '#6e4071', '#2d1e3e']
     """
-    import matplotlib as mpl
-    import matplotlib.colors as mcolors
 
-    cdict = mpl._cm.cubehelix(gamma, start, rot, hue)
-    cubehelix_cmap = mcolors.LinearSegmentedColormap("cubehelix", cdict)
+    start: int = 0
+    rotation: float = 0.4
+    gamma: float = 1.0
+    hue: float = 0.8
+    light: float = 0.85
+    dark: float = 0.15
+    reverse: bool = False
 
-    def cubehelix_palette(n):
-        values = np.linspace(light, dark, n)
-        return [mcolors.rgb2hex(cubehelix_cmap(x)) for x in values]
+    def __post_init__(self):
+        self._chmap = CubeHelixMap(
+            self.start,
+            self.rotation,
+            self.gamma,
+            self.hue,
+            self.light,
+            self.dark,
+            self.reverse,
+        )
 
-    return cubehelix_palette
+    def __call__(self, n: int) -> Sequence[RGBHexColor]:
+        return self._chmap.discrete_palette(n)
 
 
-def identity_pal():
+def identity_pal() -> Callable[[], Any]:
     """
     Create palette that maps values onto themselves
 
