@@ -1,17 +1,27 @@
 from __future__ import annotations
 
 import typing
-from collections import OrderedDict, defaultdict
-from collections.abc import Iterator
-from datetime import timezone, tzinfo
-from itertools import chain
+from datetime import datetime, timezone
+from typing import overload
 from warnings import warn
 
 import numpy as np
+import pandas.api.types as pdtypes
 
 if typing.TYPE_CHECKING:
-    from datetime import datetime
-    from typing import Sequence
+    from datetime import tzinfo
+    from typing import Any, Optional, Sequence
+
+    from mizani.typing import (
+        AnyArrayLike,
+        DurationUnit,
+        FloatArrayLike,
+        FloatVector,
+        NullType,
+        NumericUFunction,
+        SeqDatetime,
+        TupleFloat2,
+    )
 
 
 __all__ = [
@@ -19,8 +29,6 @@ __all__ = [
     "min_max",
     "match",
     "precision",
-    "first_element",
-    "multitype_sort",
     "same_log10_order_of_magnitude",
     "identity",
     "get_categories",
@@ -30,48 +38,62 @@ __all__ = [
 DISCRETE_KINDS = "ObUS"
 CONTINUOUS_KINDS = "ifuc"
 
-SECONDS = OrderedDict(
-    [
-        ("ns", 1e-9),  # nanosecond
-        ("us", 1e-6),  # microsecond
-        ("ms", 1e-3),  # millisecond
-        ("s", 1),  # second
-        ("m", 60),  # month
-        ("h", 3600),  # hour
-        ("d", 24 * 3600),  # day
-        ("w", 7 * 24 * 3600),  # week
-        ("M", 31 * 24 * 3600),  # month
-        ("y", 365 * 24 * 3600),  # year
-    ]
-)
+SECONDS: dict[DurationUnit, float] = {
+    "ns": 1e-9,  # nanosecond
+    "us": 1e-6,  # microsecond
+    "ms": 1e-3,  # millisecond
+    "s": 1,  # second
+    "m": 60,  # month
+    "h": 3600,  # hour
+    "d": 24 * 3600,  # day
+    "w": 7 * 24 * 3600,  # week
+    "M": 31 * 24 * 3600,  # month
+    "y": 365 * 24 * 3600,  # year
+}
 
-NANOSECONDS = OrderedDict(
-    [
-        ("ns", 1),  # nanosecond
-        ("us", 1e3),  # microsecond
-        ("ms", 1e6),  # millisecond
-        ("s", 1e9),  # second
-        ("m", 60e9),  # month
-        ("h", 3600e9),  # hour
-        ("d", 24 * 3600e9),  # day
-        ("w", 7 * 24 * 3600e9),  # week
-        ("M", 31 * 24 * 3600e9),  # month
-        ("y", 365 * 24 * 3600e9),  # year
-    ]
-)
+NANOSECONDS: dict[DurationUnit, float] = {
+    "ns": 1,  # nanosecond
+    "us": 1e3,  # microsecond
+    "ms": 1e6,  # millisecond
+    "s": 1e9,  # second
+    "m": 60e9,  # month
+    "h": 3600e9,  # hour
+    "d": 24 * 3600e9,  # day
+    "w": 7 * 24 * 3600e9,  # week
+    "M": 31 * 24 * 3600e9,  # month
+    "y": 365 * 24 * 3600e9,  # year
+}
 
 
-def round_any(x, accuracy, f=np.round):
+@overload
+def round_any(
+    x: FloatArrayLike, accuracy: float, f: NumericUFunction = np.round
+) -> FloatVector:
+    ...
+
+
+@overload
+def round_any(
+    x: float, accuracy: float, f: NumericUFunction = np.round
+) -> float:
+    ...
+
+
+def round_any(
+    x: FloatArrayLike | float, accuracy: float, f: NumericUFunction = np.round
+) -> FloatVector | float:
     """
     Round to multiple of any number.
     """
     if not hasattr(x, "dtype"):
         x = np.asarray(x)
 
-    return f(x / accuracy) * accuracy
+    return f(x / accuracy) * accuracy  # type: ignore
 
 
-def min_max(x, na_rm=False, finite=True):
+def min_max(
+    x: FloatArrayLike | float, na_rm: bool = False, finite: bool = True
+) -> TupleFloat2:
     """
     Return the minimum and maximum of x
 
@@ -89,8 +111,7 @@ def min_max(x, na_rm=False, finite=True):
     out : tuple
         (minimum, maximum) of x
     """
-    if not hasattr(x, "dtype"):
-        x = np.asarray(x)
+    x = np.asarray(x)
 
     if na_rm and finite:
         x = x[np.isfinite(x)]
@@ -107,50 +128,54 @@ def min_max(x, na_rm=False, finite=True):
         return float("-inf"), float("inf")
 
 
-def match(v1, v2, nomatch=-1, incomparables=None, start=0):
+def match(
+    v1: AnyArrayLike,
+    v2: AnyArrayLike,
+    nomatch: int = -1,
+    incomparables: Optional[Any] = None,
+    start: int = 0,
+) -> list[int]:
     """
     Return a vector of the positions of (first)
     matches of its first argument in its second.
 
     Parameters
     ----------
-    v1: array_like
-        Values to be matched
+    v1: array-like
+        The values to be matched
 
-    v2: array_like
-        Values to be matched against
+    v2: array-like
+        The values to be matched against
 
     nomatch: int
-        Value to be returned in the case when
+        The value to be returned in the case when
         no match is found.
 
-    incomparables: array_like
-        Values that cannot be matched. Any value in ``v1``
-        matching a value in this list is assigned the nomatch
-        value.
+    incomparables: array-like
+        A list of values that cannot be matched.
+        Any value in v1 matching a value in this list
+        is assigned the nomatch value.
     start: int
         Type of indexing to use. Most likely 0 or 1
     """
-    v2_indices = {}
-    for i, x in enumerate(v2):
-        if x not in v2_indices:
-            v2_indices[x] = i
+    # NOTE: This function gets called a lot. If it can
+    # be optimised, it should.
+    lookup: dict[Any, int] = {}
+    for i, x in enumerate(v2, start=start):
+        if x not in lookup:
+            lookup[x] = i
 
-    v1_to_v2_map = [nomatch] * len(v1)
-    skip = set(incomparables) if incomparables else set()
-    for i, x in enumerate(v1):
-        if x in skip:
-            continue
-
-        try:
-            v1_to_v2_map[i] = v2_indices[x] + start
-        except KeyError:
-            pass
-
-    return v1_to_v2_map
+    if incomparables:
+        skip = set(incomparables)
+        lst = [
+            lookup.get(x, nomatch) if x not in skip else nomatch for x in v1
+        ]
+    else:
+        lst = [lookup.get(x, nomatch) for x in v1]
+    return lst
 
 
-def precision(x):
+def precision(x: FloatArrayLike | float) -> float:
     """
     Return the precision of x
 
@@ -189,59 +214,6 @@ def precision(x):
         return 1
     else:
         return 10 ** int(np.floor(np.log10(span)))
-
-
-def first_element(obj):
-    """
-    Return the first element of `obj`
-
-    Parameters
-    ----------
-    obj : iterable
-        Should not be an iterator
-
-    Returns
-    -------
-    out : object
-        First element of `obj`. Raise a class:`StopIteration`
-        exception if `obj` is empty.
-    """
-    if isinstance(obj, Iterator):
-        raise RuntimeError("Cannot get the first element of an iterator")
-    return next(iter(obj))
-
-
-def multitype_sort(a):
-    """
-    Sort elements of multiple types
-
-    x is assumed to contain elements of different types, such that
-    plain sort would raise a `TypeError`.
-
-    Parameters
-    ----------
-    a : array-like
-        Array of items to be sorted
-
-    Returns
-    -------
-    out : list
-        Items sorted within their type groups.
-    """
-    types = defaultdict(list)
-    numbers = {int, float, complex}
-
-    for x in a:
-        t = type(x)
-        if t in numbers:
-            types["number"].append(x)
-        else:
-            types[t].append(x)
-
-    for t in types:
-        types[t] = np.sort(types[t])
-
-    return list(chain.from_iterable(types[t] for t in types))
 
 
 def same_log10_order_of_magnitude(x, delta=0.1):
@@ -325,23 +297,29 @@ def log(x, base):
     return res
 
 
-def get_timezone(x: Sequence[datetime]) -> tzinfo:
+def get_timezone(x: SeqDatetime) -> tzinfo | None:
     """
     Return a single timezone for the sequence of datetimes
 
-    Returns the timezone of first item and warns if any other item
-    has a different timezone
+    Returns the timezone of first item and warns if any other items
+    have a different timezone
     """
 
     # Ref: https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+    x0 = next(iter(x))
+    if not isinstance(x0, datetime):
+        return None
 
-    info = x[0].tzinfo
+    info = x0.tzinfo
     if info is None:
         return timezone.utc
 
     # Consistency check
-    tzname0 = info.tzname(x[0])
-    tznames = (dt.tzinfo.tzname(dt) if dt.tzinfo else None for dt in x)
+    tzname0 = info.tzname(x0)
+    tznames = (
+        dt.tzinfo.tzname(dt) if dt.tzinfo else None for dt in x  # type: ignore
+    )
+
     if any(tzname0 != name for name in tznames):
         msg = (
             "Dates in column have different time zones. "
@@ -351,3 +329,32 @@ def get_timezone(x: Sequence[datetime]) -> tzinfo:
         )
         warn(msg.format(tzname0))
     return info
+
+
+def get_null_value(x: Any) -> NullType:
+    """
+    Return a Null value for the type of values
+    """
+    from datetime import datetime, timedelta
+
+    import pandas as pd
+
+    x0 = next(iter(x))
+    numeric_types: Sequence[type] = (np.int64, np.float64, int, float, bool)
+
+    if pdtypes.is_object_dtype(x):
+        return None
+    elif isinstance(x0, numeric_types):  # type: ignore
+        return float("nan")
+    # pandas types subclass cypthon types, so check
+    # for them first
+    elif isinstance(x0, (pd.Timestamp, pd.Timedelta)):
+        return type(x0)("NaT")
+    elif isinstance(x0, (datetime, timedelta)):
+        return None
+    elif isinstance(x0, (np.datetime64, np.timedelta64)):
+        return type(x0)("NaT")
+    else:
+        raise ValueError(
+            "Cannot get a null value for type: {}".format(type(x[0]))
+        )
